Add initial mechanism for transforming Media events to RemotePlayer
